chore(level): Move signle agent to a different file
